Python: share the durable entity factory and workflow context across hosts
Extracts the configured agent entity class into create_agent_entity_class so the DurableTask worker and any other host register the same implementation, instead of the worker keeping a private copy.

Also makes DurableTaskWorkflowContext take supports_event_streaming so a host that cannot publish the accumulating event log can opt out. Defaults to True, so behaviour is unchanged.
