+++ python/packages/durabletask/agent_framework_durabletask/_worker.py
@@ -16,9 +16,8 @@ from agent_framework import SupportsAgentRun, Workflow
 from durabletask.task import ActivityContext, OrchestrationContext
 from durabletask.worker import TaskHubGrpcWorker
 
-from ._async_bridge import run_agent_coroutine
 from ._callbacks import AgentResponseCallbackProtocol
-from ._entities import AgentEntity, DurableTaskEntityStateProvider
+from ._entities import create_agent_entity_class
 from ._workflows.activity import execute_workflow_activity
 from ._workflows.dt_context import DurableTaskWorkflowContext
 from ._workflows.naming import (
@@ -137,7 +136,7 @@ class DurableAIAgentWorker:
         effective_callback = callback or self._callback
 
         # Create a configured entity class using the factory
-        entity_class = self.__create_agent_entity(agent, effective_callback, entity_id=registration_name)
+        entity_class = create_agent_entity_class(agent, effective_callback, entity_id=registration_name)
 
         # Register the entity class with the worker
         # The worker.add_entity method takes a class
@@ -349,73 +348,3 @@ class DurableAIAgentWorker:
 
         self._worker.add_orchestrator(workflow_orchestrator)
         logger.debug("[DurableAIAgentWorker] Registered workflow orchestrator: %s", orchestrator_name)
-
-    def __create_agent_entity(
-        self,
-        agent: SupportsAgentRun,
-        callback: AgentResponseCallbackProtocol | None = None,
-        *,
-        entity_id: str | None = None,
-    ) -> type[DurableTaskEntityStateProvider]:
-        """Factory function to create a DurableEntity class configured with an agent.
-
-        This factory creates a new class that combines the entity state provider
-        with the agent execution logic. Each agent gets its own entity class.
-
-        Args:
-            agent: The agent instance to wrap
-            callback: Optional callback for agent responses
-            entity_id: Optional identity to register the entity under instead of
-                ``agent.name`` (used by workflow hosting to key entities by
-                executor id).
-
-        Returns:
-            A new DurableEntity subclass configured for this agent
-        """
-        agent_name = entity_id or agent.name or type(agent).__name__
-        entity_name = f"dafx-{agent_name}"
-
-        class ConfiguredAgentEntity(DurableTaskEntityStateProvider):
-            """Durable entity configured with a specific agent instance."""
-
-            def __init__(self) -> None:
-                super().__init__()
-                # Create the AgentEntity with this state provider
-                self._agent_entity = AgentEntity(
-                    agent=agent,
-                    callback=callback,
-                    state_provider=self,
-                )
-                logger.debug(
-                    "[ConfiguredAgentEntity] Initialized entity for agent: %s (entity name: %s)",
-                    agent_name,
-                    entity_name,
-                )
-
-            def run(self, request: Any) -> Any:
-                """Handle run requests from clients or orchestrations.
-
-                Args:
-                    request: RunRequest as dict or string
-
-                Returns:
-                    AgentResponse as dict
-                """
-                logger.debug("[ConfiguredAgentEntity.run] Executing agent: %s", agent_name)
-                # Run on the shared persistent loop so async resources created by
-                # shared agent clients/credentials stay bound to a live loop across
-                # successive entity invocations (avoids cross-loop hangs).
-                response = run_agent_coroutine(self._agent_entity.run(request))
-                return response.to_dict()
-
-            def reset(self) -> None:
-                """Reset the agent's conversation history."""
-                logger.debug("[ConfiguredAgentEntity.reset] Resetting agent: %s", agent_name)
-                self._agent_entity.reset()
-
-        # Set the entity name to match the prefixed agent name
-        # This is used by durabletask to register the entity
-        ConfiguredAgentEntity.__name__ = entity_name
-        ConfiguredAgentEntity.__qualname__ = entity_name
-
-        return ConfiguredAgentEntity
